Method for getting price coef moved from metro parser to init.
