+++ backend/providers/__init__.py
@@ -20,6 +20,17 @@ def extract_price(price_text):
     return price_in_number
 
 
+def get_price_coefficient(title):
+    weight_number, weight_measure = re.findall(r"\b(\d+,?\d*)(г|кг)\b", title)[0]
+    weight_number = re.sub(" ", "", weight_number.strip())
+    weight_number = float(re.sub(",", ".", weight_number))
+    if weight_measure == "г":
+        price_coef = 1 / (weight_number / 1000)
+    else:
+        price_coef = 1 / weight_number
+    return price_coef
+
+
 def clean_title(title, producer):
     title = re.sub(r"\b(\d+)(г|кг)\b", "", title.lower())
     title = re.sub(producer, "", title).strip()
